requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
passlib==1.7.4
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Cors==4.0.0
factory-boy==3.3.0